Dynamic rate change bug fix
When the design boots at 25G and the user uses ethtool to change speed to 10G
it works great however, returning back to 25G was not possible due to missing
protocol configs in 3 lanes. Please check AR000037047 - Issue with dynamic rate switching on VCK190-Ethernet-TRD - MRMAC design
for more info.
https://adaptivesupport.amd.com/s/article/000037047?language=en_US